04_serverless_stream_qna_rag/requirements.txt
aws-cdk-lib==2.90.0
constructs>=10.0.0,<11.0.0
sagemaker
boto3
python-dotenv==1.0.0